# Conflicts:
#	.idea/deploymentTargetSelector.xml
#	app/src/main/java/com/swen549/touchanalytics/data/FirebaseClient.kt

+++ app/src/main/java/com/swen549/touchanalytics/data/TouchalyticsApiService.kt
@@ -9,7 +9,7 @@ import retrofit2.http.Path
 interface TouchalyticsApiService {
     @POST("authenticate/{userID}")
     suspend fun authenticate(
-        @Path("userID") userID: Int,
+        @Path("userID") userID: Long,
         @Body feature: Feature
     ): Response<JsonObject>
 }
